Some graphics added to bullet and background, solved factory problem that creates an alien with null value_

+++ src/spaceinvadders/Bullet.java
@@ -4,23 +4,28 @@ import org.academiadecodigo.simplegraphics.pictures.Picture;
 
 public class Bullet {
 
-    Picture representation;
+    private Picture representation;
+    private Shootable[] targets;
 
-    public Bullet(Character character) {
-        if (character instanceof SpaceShip) {
-            representation = new Picture(character.getX() + (character.getWidth() / 2), character.getY() - character.getHeight() / 2, "/Users/albertoreis/dev/spaceinvadersgroup/resources/projectile.png");
+
+    public Bullet(Shootable shootable) {
+        int centerX = shootable.getX() + (shootable.getWidth() / 2);
+        int centerY = shootable.getY() - (shootable.getHeight() / 2);
+
+        if (shootable instanceof SpaceShip) {
+            representation = new Picture(centerX, centerY, "/Users/albertoreis/dev/spaceinvadersgroup/resources/projectile.png");
         }
-        if (character instanceof Boss) {
-            representation = new Picture(character.getX() + (character.getWidth() / 2), character.getY() + character.getHeight(), "/Users/albertoreis/dev/spaceinvadersgroup/resources/bossprojectile.png");
+        if (shootable instanceof Boss) {
+            representation = new Picture(centerX, centerY, "/Users/albertoreis/dev/spaceinvadersgroup/resources/bossprojectile.png");
         }
-        if (character instanceof Alien){
-            representation = new Picture(character.getX() + (character.getWidth() / 2), character.getY() + character.getHeight(), "/Users/albertoreis/dev/spaceinvadersgroup/resources/enemyprojectile.png");
+        if (shootable instanceof Alien) {
+            representation = new Picture(centerX, centerY, "/Users/albertoreis/dev/spaceinvadersgroup/resources/enemyprojectile.png");
         }
+        representation.draw();
     }
 
-    public void shootUpwards(Character[] characters) {
-        representation.draw();
-        Character[] character = characters;
+    public void shootUpwards(Shootable[] shootables) {
+        targets = shootables;
 
         Thread t1 = new Thread(new Runnable() {
             public void run() {
@@ -31,31 +36,43 @@ public class Bullet {
                         Thread.currentThread().interrupt();
                     }
                     representation.translate(0, -3);
+                    //System.out.println(Arrays.toString(targets));
+                    for (int i = targets.length - 1; i >= 0; i--) {
+                        System.out.println("inicio for indice: " + i);
+                        if (targets[i].isVisible() && isSamePos(targets[i])) {
+                            representation.delete();
+                            targets[i].hit();
+                            System.out.println("aaaa");
+                        }
+                        //if (hitChecker(targets[i])) {
+                        //  System.out.println("Aqui! " + targets[i]);
+                        //}
+                    }
                 }
                 representation.delete();
             }
         });
         t1.start();
-        for (int i = characters.length - 1; i >= 0; i--) {
-            if (hitChecker(characters[i])) ;
-            break;
-        }
     }
 
-    public boolean hitChecker(Character target) {
-        if (!target.isDead()) {
-            if (target instanceof Alien) {
-                if (representation.getY() == target.getY() + target.getHeight()
-                        && representation.getX() <= target.getX() + target.getWidth()
-                        && representation.getX() >= target.getX()) {
-                    representation.delete();
-                    target.kill();
-                    Explosion.explode(target);
-                    return true;
-                }
-            }
-        }return false;
+    private boolean isSamePos(Shootable target) {
+        return representation.getY() == target.getY() + target.getHeight()
+                && representation.getX() <= target.getX() + target.getWidth()
+                && representation.getX() >= target.getX();
     }
+
+    public boolean hitChecker(Shootable target) {
+        if (target.isVisible()
+                && representation.getY() == target.getY() + target.getHeight()
+                && representation.getX() <= target.getX() + target.getWidth()
+                && representation.getX() >= target.getX()) {
+            representation.delete();
+            target.hit();
+            return true;
+        }
+        return false;
+    }
+
 
     public void shootDownwards() {
         representation.draw();
